Split up import manager to be more manageable (this is slightly out of date with researcher.py CSV validation/deduplication changes)
